Mend unsigned to double conversion.
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@22408 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/test/TestNonCachedTetrahedralMesh.hpp
@@ -103,7 +103,7 @@ public:
             mesh.ConstructCuboid(30,30,30);
             non_cached_mem_usage = GetMemoryUsage();
         }
-        unsigned non_cached_construction_time = Timer::GetElapsedTime();
+        double non_cached_construction_time = Timer::GetElapsedTime();
 
         // Constructing the non cached object should be quicker
         // Note: this does occasionally fail, due to other activity on the machine
